config file added, trying to track the view time per clip
