+++ src/configuration.js
@@ -1,0 +1,11 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyA4J2CKgtar9j_04i4y6NnsO5YqsR2Wnio",
+    authDomain: "harry-c24d2.firebaseapp.com",
+    projectId: "harry-c24d2",
+    storageBucket: "harry-c24d2.appspot.com",
+    messagingSenderId: "689249194788",
+    appId: "1:689249194788:web:546cc4d4dfc50e9c4dd9b1",
+    measurementId: "G-Q6TQGHPG1X" 
+  };
+
+  export default firebaseConfig
